Support Legacy Meteor which expects Node on port 80
Use iptables to forward port 80 at the kernel level to the node port (3000).  This allows us to run Node as unprivileged and still be on port 80.

My preferred, but not possible, alternative to this was either:

Option 1: Using the kernel `setcap CAP_NET_BIND_SERVICE=+eip1` which allows unprivileged users to bind to low-numbered ports.  Unfortunately, this requires support by the filesystem as the attribute is stored on the binary.  Since CircleCI uses an Ubuntu without the right kernel for aufs, they use btrfs which doesn't support this.  Blast.
Option 2: Start the docker container with --cap-add=NET_BIND_SERVICE.  Ok, great! But again, I have no control over the way that Meteor Up starts and it doesn't currently do that.

So here we are.