Add babel-runtime for all apps >= 1.4
Due to the changes in Meteor 1.4.2.1 for babel-runtime and the associated bump in version of Meteor's `babel-runtime` package to 1.0.0 (which no longer provides `babel-runtime` functionality itself) it is now necessary to `meteor npm install` the `babel-runtime` package for all newly created 1.4 apps.  It is necessary for all 1.4 apps (not just 1.4.2.1 or higher) because of the general Atmosphere-package version-unpinning as of the 1.4 release.